6_ft_check_in_cluster..

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: spark <spark@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/29 17:02:05 by spark             #+#    #+#             */
-/*   Updated: 2020/10/05 15:46:45 by spark            ###   ########.fr       */
+/*   Updated: 2020/10/05 20:37:25 by spark            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
